converted to using TS pipeline
