+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 
 examples/**/*.js
 build/
+dist/
